新增仪表盘控件

+++ app/src/main/java/com/example/wangweimin/customerview/view/MyPieChart.java
@@ -174,7 +174,7 @@ public class MyPieChart extends View {
         }
     }
 
-    private float convertDp2Px(float dp) {
+    private float convertDp2Px(int dp) {
         Resources r = getResources();
         return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_PX, dp, r.getDisplayMetrics());
     }
